# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Blog backend API",
        "title": "z-blog-backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/article": {
            "get": {
                "description": "分页获取文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "获取文章列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/article.ArticleListByPageResponse"
                        }
                    }
                }
            }
        },
        "/article/hot_articles": {
            "get": {
                "description": "获取热门文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "获取热门文章",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/article.ArticleListResponse"
                        }
                    }
                }
            }
        },
        "/article/{id}": {
            "get": {
                "description": "根据id获取文章详情（带正文）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "获取文章详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/article.ArticleDetailResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "article.Article": {
            "type": "object",
            "properties": {
                "authorName": {
                    "description": "作者",
                    "type": "string"
                },
                "content": {
                    "description": "正文",
                    "type": "string"
                },
                "cover": {
                    "description": "封面",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "desc": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_delete": {
                    "type": "boolean"
                },
                "status": {
                    "description": "状态",
                    "type": "integer"
                },
                "tags": {
                    "description": "标签（逗号分隔形式）",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "views": {
                    "description": "浏览数",
                    "type": "integer"
                }
            }
        },
        "article.ArticleDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/article.Article"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "article.ArticleListByPageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object",
                    "properties": {
                        "data": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/article.ArticleWithoutContent"
                            }
                        },
                        "page": {
                            "type": "integer"
                        },
                        "size": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "article.ArticleListResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/article.ArticleWithoutContent"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "article.ArticleWithoutContent": {
            "type": "object",
            "properties": {
                "authorName": {
                    "description": "作者",
                    "type": "string"
                },
                "cover": {
                    "description": "封面",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "tags": {
                    "description": "标签（逗号分隔）",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "views": {
                    "description": "浏览数",
                    "type": "integer"
                }
            }
        }
    }
}